feat(verification): 新增邮箱验证功能，重构相关处理逻辑
